requirements.txt
ultralytics
fastapi
uvicorn
python-multipart
streamlit
requests
pillow
opencv-python